Set cm:ownable aspect with cm:owner set to "admin" on case nodes. This ensures that the cm:creator does not get full control.
Set case members edit/delete webscript to run as normal user, not admin.

Make CaseSimpleWriter be a Collaborator instead of Coordinator (so it cannot have full control).

Do not inherit parent permissions for a case node when setting up. This means that it will not Allow GROUP_EVERYONE to have Read permission.

Setup case permission groups in the Share zone, so they don't appear in the UI.

Setup case permission groups so that only case owners can modify the role groups (not only admin as before).